requirements.txt
streamlit
google-generativeai
faiss-cpu
python-dotenv
numpy
